Move segment_storage.py into segment.py and rename classes to avoid confusion with file-based storage
